address kv_cache_reuse
