+++ tensorrt_llm/inputs/registry.py
@@ -1,10 +1,13 @@
 import enum
+import random
 from dataclasses import dataclass, field
 from typing import (Any, Callable, Dict, List, Optional, Protocol, Tuple, Type,
                     TypeVar)
 
 from PIL import Image
 from torch import Tensor, nn
+
+import tensorrt_llm
 
 from .._utils import nvtx_range_debug
 from ..logger import logger
@@ -47,9 +50,41 @@ class BaseDummyInputsBuilder:
     Base class for generating dummy inputs. Specially for profiling
     """
 
+    def __init__(self, **kwargs):
+        super().__init__(**kwargs)
+        self.image_max_dim = 16384
+        self.img_min_dim = 128
+
+    def get_dummy_image(self, max_width: int, max_height: int):
+        image = Image.new("RGB", (max_width, max_height),
+                          color=random.randint(0, 256))
+        return image
+
     def get_dummy_prompt(self, input_seq_len: int):
-        raise NotImplementedError(
-            "Please ensure this method is implemented in your inherited class")
+        # we use the max resolution as starting point
+        # img_max_dim = 9999  # TODO: Find better starting resolution
+        while self.image_max_dim >= self.img_min_dim:
+            image = self.get_dummy_image(max_width=self.image_max_dim,
+                                         max_height=self.image_max_dim)
+
+            test_mm_prompt = tensorrt_llm.inputs.utils.default_multimodal_input_loader(
+                tokenizer=self.tokenizer,
+                model_dir=self.model_path,
+                model_type=self.model_config.model_type,
+                modality="image",
+                prompts=[""],
+                media=[[image]],
+                image_data_format="pt")[0]
+
+            prompt_token_ids_single_img, _ = self(test_mm_prompt, None)
+
+            if len(prompt_token_ids_single_img) <= input_seq_len:
+                return test_mm_prompt
+
+            # reduce img resolution
+            self.image_max_dim = self.image_max_dim >> 1
+
+        return None
 
 
 class BaseMultimodalInputProcessor:
@@ -60,6 +95,9 @@ class BaseMultimodalInputProcessor:
     This class provides default implementations that work with most AutoProcessor-based
     models. Specific processors can override these methods if they need custom logic.
     """
+
+    def __init__(self, **kwargs):
+        super().__init__(**kwargs)
 
     def get_processor(self) -> Optional[Any]:
         """Return the processor object if available; otherwise raise NotImplementedError.
